First Commit of 16 Working Structural Metrics
Halstead Length is the sum of the total number of operators and operand.

Halstead Vocabulary is the sum of the number of unique operators and
unique operands.

Halstead Volume is the program length times the log2 of the program
vocabulary.

Halstead Difficulty is half of the unique operators multiplied by the
total number of operands, divided by the number of distinct operators.

Halstead Effort is the difficulty multiplied by the volume. Effort = DV.
Effort was intended as a suggestion for how long code review might take.

Number of comments.

Number of lines of comments.

Number of looping statements.

Number of operators.

Number of operands.

Number of expressions.

Number of variable declarations.

External Method references are the number of methods called from an
external class.

Local method references are the number of methods called from the same
class.

Number of casts.

Maintainability Index measures how maintainable the source code is.